test(typechecker): refresh parity review gates
