+++ crates/vize_canon/tests/component_event_arity.rs
@@ -152,7 +152,8 @@ const item = {} as {
         app.content
     );
     assert!(
-        app.content.contains("if (__vize_handler_"),
+        app.content
+            .contains("if (typeof __vize_handler_7_15 === \"function\")"),
         "optional event references should be guarded before invocation:\n{}",
         app.content
     );
